requirements.txt
pillow
image
colorama